# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "E-commerce API",
    "description": "<h2>To be able to bypass the JWT authentication for the endpoints needing authorization,         login first, then use the token recieved, press the <b>Authorize</b> button and type 'Bearer ' and paste the token copied.         After, you can access the enpoints requiring authorization without filling the <b>Authorization</b> field.</h2>"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization"
    }
  },
  "paths": {
...
    "/brands/": {
      "get": {
        "tags": [
          "The Brand Table Endpoints"
        ],
        "description": "<h3>Gets the brand list.</h3>",
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "post": {
        "tags": [
          "The Brand Table Endpoints"
        ],
        "description": "<h3>If user is authorized and admin, creates a brand based on validated parameters provided in the request's body.</h3>",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/brandCreate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "The Brand Table Endpoints"
        ],
        "description": "<h3>If user is authorized and admin, changes a brand based on validated parameters provided in the request's body.</h3>",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/brandChange"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "The Brand Table Endpoints"
        ],
        "description": "<h3>If user is authorized and admin, removes a brand based on validated parameters provided in the request's body. This happens only if there are no products with this brand.</h3>",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/brandDelete"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/cart/": {
      "get": {
        "tags": [
          "The Cart & Cart Item Tables Endpoints"
        ],
        "description": "<h3>Gets aunthenticated user's active cart with all the cart items in it.</h3>",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "The Cart & Cart Item Tables Endpoints"
        ],
        "description": "<h3>If user is authorized, adds a cart item to their cart based on validated parameters provided in the request's body. Also, calculates the total cost based on quantity and discount, checks for duplicate items if so updates quantity accordingly, depending on the product quantity availability makes quantitative decisions.</h3>",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/cartItemCreate"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "The Cart & Cart Item Tables Endpoints"
        ],
        "description": "<h3>If user is authorized, makes changes to cart item's quantity in their active cart based on validated parameters provided in the request's body. Also, checks if the cart is the active one, makes decisions based on available product quantity, recalculates the total cost based on quantity and discount.</h3>",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/cartItemQuantiyChange"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "The Cart & Cart Item Tables Endpoints"
        ],
        "description": "<h3>If user is authorized, removes a cart item from their active cart based on validated parameters provided in the request's body. Also, checks if the cart is the active one, recalculates the total cost based on quantity and discount.</h3>",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/cartItemDelete"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/cart/checkout/now": {
      "post": {
        "tags": [
          "The Cart & Cart Item Tables Endpoints"
        ],
        "description": "<h3>If user is authorized, checks out a cart with its items based on validated parameters provided in the request's body. Also, if there are no items inside the active cart cancels the operation, checks again if enough product availability is present, recalculates the total cost based on quantity and discount before the check-out, creates unique order number, creates order and order items, updates users membership status, soft deletes the cart with its items, creates a new cart, updates product quantity and much more...</h3>",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/admin/brands": {
      "get": {
        "tags": [
          "The Admin Front-End Endpoints"
        ],
        "description": "<h3>Renders the brands page if the token owner has an admin role. If not admin then returns the login page back to the user.</h3>",
        "produces": [
          "text/html"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "brandCreate": {
      "type": "object",
      "properties": {
        "brand": {
          "type": "string",
          "example": "Apple"
        }
      },
      "required": [
        "brand"
      ]
    },
    "brandChange": {
      "type": "object",
      "properties": {
        "brandid": {
          "type": "number",
          "example": 2
        },
        "brand": {
          "type": "string",
          "example": "Apple"
        }
      },
      "required": [
        "brandid",
        "brand"
      ]
    },
    "brandDelete": {
      "type": "object",
      "properties": {
        "brandid": {
          "type": "number",
          "example": 2
        }
      },
      "required": [
        "brandid"
      ]
    },
...
    "cartItemCreate": {
      "type": "object",
      "properties": {
        "productid": {
          "type": "number",
          "example": 1
        },
        "quantity": {
          "type": "number",
          "example": 1
        }
      },
      "required": [
        "productid",
        "quantity"
      ]
    },
    "cartItemQuantiyChange": {
      "type": "object",
      "properties": {
        "cartitemid": {
          "type": "number",
          "example": 1
        },
        "quantity": {
          "type": "number",
          "example": 1
        }
      },
      "required": [
        "cartitemid",
        "quantity"
      ]
    },
    "cartItemDelete": {
      "type": "object",
      "properties": {
        "cartitemid": {
          "type": "number",
          "example": 1
        }
      },
      "required": [
        "cartitemid"
      ]
    },
...
  }
}